+++ controllers/workspace.js
@@ -9,14 +9,20 @@ import joiWorkspaceSchema from "../joiSchemas/workspaceSchema.js";
 // Desc     :   Get All Workspaces
 
 export const getAllWorkspaces = async (req, res) => {
-	try {
-		const workspaces = await Workspaces.find({});
-		// Return a success response
-		res.status(200).json({ success: true, error: false, workspaces });
-	} catch (error) {
-		console.error(error.message);
-		res.status(400).json("Server Error");
-	}
+  try {
+    const loggedInUserId = req.userId;
+
+    const workspaces = await Workspaces.find({ creator: loggedInUserId });
+    if (workspaces) {
+      // Return a success response
+      res.status(200).json({ success: true, error: false, workspaces });
+    } else {
+      return res.status(404).json({ message: "Workspace not found." });
+    }
+  } catch (error) {
+    console.error(error.message);
+    res.status(400).json("Server Error");
+  }
 };
 
 // Path     :   /api/workspace/addWorkspace
@@ -25,32 +31,34 @@ export const getAllWorkspaces = async (req, res) => {
 // Desc     :   Create new Workspace
 
 export const addWorkspace = async (req, res) => {
-	try {
-		const workspaceData = req.body;
-		console.log(workspaceData,req.userId)
-		const { error, value } = joiWorkspaceSchema.validate(workspaceData, { abortEarly: false });
-	  
-		if (error) {
-		  const errorMessage = error.details.map((detail) => detail.message);
-		  return res.status(400).json({ success: false, error: errorMessage });
-		}
-		const loggedInUserId = req.userId;
-		const newWorkspace = await new Workspaces({
-			name: workspaceData.name,
-			creator : loggedInUserId,
-		}).save();
-		// Return a success response
-		return res.status(200).json({
-			success: true,
-			error: false,
-			message: "Workspace Added Successfully",
-			newWorkspace,
-		});
-	} catch (error) {
-		// Return an error response if an error occurs
-		console.error(error.message);
-		res.status(400).json("Server Error");
-	}
+  try {
+    const workspaceData = req.body;
+    console.log(workspaceData, req.userId);
+    const { error, value } = joiWorkspaceSchema.validate(workspaceData, {
+      abortEarly: false,
+    });
+
+    if (error) {
+      const errorMessage = error.details.map((detail) => detail.message);
+      return res.status(400).json({ success: false, error: errorMessage });
+    }
+    const loggedInUserId = req.userId;
+    const newWorkspace = await new Workspaces({
+      name: workspaceData.name,
+      creator: loggedInUserId,
+    }).save();
+    // Return a success response
+    return res.status(200).json({
+      success: true,
+      error: false,
+      message: "Workspace Added Successfully",
+      newWorkspace,
+    });
+  } catch (error) {
+    // Return an error response if an error occurs
+    console.error(error.message);
+    res.status(400).json("Server Error");
+  }
 };
 
 // Path     :   /api/workspace/update-workspace/:id
@@ -58,28 +66,35 @@ export const addWorkspace = async (req, res) => {
 // Access   :   Private
 // Desc     :   Update Workspace
 export const updateWorkspace = async (req, res) => {
-	const workspaceId = req.params.id;
-	console.log(req.params.id);
-	const { name } = req.body;
+  const workspaceId = req.params.id;
+  const loggedInUserId = req.userId;
 
-	try {
-		let workspaceData = await Workspaces.findOne({ _id: workspaceId });
-		console.log(workspaceData);
+  console.log(req.params.id);
+  const { name } = req.body;
 
-		if (!workspaceData) {
-			return res.status(404).json({ message: "Workspace not found." });
-		}
-		(workspaceData.name = name), await workspaceData.save();
+  try {
+    let workspaceData = await Workspaces.findOne({ _id: workspaceId });
 
-		// Return a success response
-		return res
-			.status(200)
-			.json({ workspaceData, message: "Workspace updated successfully." });
-	} catch (error) {
-		// Return an error response if an error occurs
-		console.error(error);
-		return res.json({ message: "Server error." });
-	}
+    console.log(workspaceData);
+
+    if (!workspaceData) {
+      return res.status(404).json({ message: "Workspace not found." });
+    }
+
+    if ((workspaceData.creator = loggedInUserId)) {
+      (workspaceData.name = name), await workspaceData.save();
+      // Return a success response
+      return res
+        .status(200)
+        .json({ workspaceData, message: "Workspace updated successfully." });
+    } else {
+      return res.status(401).json({ message: "Unauthorized User" });
+    }
+  } catch (error) {
+    // Return an error response if an error occurs
+    console.error(error);
+    return res.json({ message: "Server error." });
+  }
 };
 
 // Path     :   /api/workspace/delete-workspace/:id
@@ -88,23 +103,28 @@ export const updateWorkspace = async (req, res) => {
 // Desc     :   Delete Workspace
 
 export const deleteWorkspace = async (req, res) => {
-	try {
-		const id = req.params.id;
-		console.log(id);
+  try {
+    const id = req.params.id;
+    const loggedInUserId = req.userId;
+    console.log(id);
 
-		let workspace = await Workspaces.findOne({ _id: id });
-		console.log(workspace);
-		if (!workspace) {
-			return res.status(404).json({ message: "Workspace not found" });
-		}
-		await workspace.deleteOne();
-		// Return a success response
-		res.status(200).json({ message: "Workspace deleted successfully" });
-	} catch (err) {
-		// Return an error response if an error occurs
-		console.error(err);
-		res.json({ message: "Server error" });
-	}
+    let workspace = await Workspaces.findOne({ _id: id });
+    console.log(workspace);
+    if (!workspace) {
+      return res.status(404).json({ message: "Workspace not found" });
+    }
+    if ((workspace.creator = loggedInUserId)) {
+      await workspace.deleteOne();
+      // Return a success response
+      res.status(200).json({ message: "Workspace deleted successfully" });
+    } else {
+      return res.status(401).json({ message: "Unauthorized User" });
+    }
+  } catch (err) {
+    // Return an error response if an error occurs
+    console.error(err);
+    res.json({ message: "Server error" });
+  }
 };
 
 // Path     :   /api/workspace/add-collaborators/:workspaceId
@@ -112,64 +132,70 @@ export const deleteWorkspace = async (req, res) => {
 // Access   :   Private
 // Desc     :   Add Collaborator
 export const addCollaborator = async (req, res) => {
-	const workspaceId = req.params.id;
-	console.log(req.params.id);
-	const { email } = req.body;
-	const collaborator = await User.findOne({ email: email });
-	console.log(collaborator)
-	if (!collaborator) {
-		return res.status(404).json({ message: "Send Registration link" });
-	} else {
-		try {
-			let workspaceData = await Workspaces.findOne({ _id: workspaceId });
-			console.log(workspaceData);
+  const workspaceId = req.params.id;
+  const loggedInUserId = req.userId;
 
-			if (!workspaceData) {
-				return res.status(404).json({ message: "Workspace not found." });
-			}
-			const newCollaborators = [
-				...workspaceData.collaborators,
-				collaborator.id,
-			];
-			const addCol = await Workspaces.findByIdAndUpdate(
-				workspaceId,
-				{ collaborators: newCollaborators },
-				{
-					new: true,
-				}
-			);
-			await sendEmailToCollaborator(collaborator.email);
-			// Return a success response
-			return res
-				.status(200)
-				.json({ addCol, message: "Collaborator added successfully." });
-		} catch (error) {
-			// Return an error response if an error occurs
-			console.error(error);
-			return res.status(500).json({ message: "Server error." });
-		}
-	}
+  console.log(req.params.id);
+  const { email } = req.body;
+  const collaborator = await User.findOne({ email: email });
+  console.log(collaborator);
+  if (!collaborator) {
+    return res.status(404).json({ message: "Send Registration link" });
+  } else {
+    try {
+      let workspaceData = await Workspaces.findOne({ _id: workspaceId });
+      console.log(workspaceData);
+
+      if (!workspaceData) {
+        return res.status(404).json({ message: "Workspace not found." });
+      }
+      if ((workspaceData.creator = loggedInUserId)) {
+        const newCollaborators = [
+          ...workspaceData.collaborators,
+          collaborator.id,
+        ];
+        const addCol = await Workspaces.findByIdAndUpdate(
+          workspaceId,
+          { collaborators: newCollaborators },
+          {
+            new: true,
+          }
+        );
+        await sendEmailToCollaborator(collaborator.email);
+        // Return a success response
+        return res
+          .status(200)
+          .json({ addCol, message: "Collaborator added successfully." });
+      } else {
+        return res.status(401).json({ message: "Unauthorized User" });
+      }
+    } catch (error) {
+      // Return an error response if an error occurs
+      console.error(error);
+      return res.status(500).json({ message: "Server error." });
+    }
+  }
 };
 // Send Email to collaborator
 const sendEmailToCollaborator = async (to) => {
-	console.log(to);
-	try {
-		const transporter = nodemailer.createTransport({
-			host: "smtp.ethereal.email",
-			port: 587,
-			auth: {
-				user: "austen.denesik@ethereal.email",
-				pass: "Mat42VrqwEztUD9Hm3",
-			},
-		});
-		await transporter.sendMail({
-			from: "admin@gmail.com",
-			to: to,
-			subject: "Welcome to E-NoteHub",
-			html: `<p>Hi! I am adding you as collaborator in my workspace `,
-		});
-		console.log("Email sent");
-	} catch (error) {
-		console.log("Email not sent", error);
-	}
+  console.log(to);
+  try {
+    const transporter = nodemailer.createTransport({
+      host: "smtp.ethereal.email",
+      port: 587,
+      auth: {
+        user: "austen.denesik@ethereal.email",
+        pass: "Mat42VrqwEztUD9Hm3",
+      },
+    });
+    await transporter.sendMail({
+      from: "admin@gmail.com",
+      to: to,
+      subject: "Welcome to E-NoteHub",
+      html: `<p>Hi! I am adding you as collaborator in my workspace `,
+    });
+    console.log("Email sent");
+  } catch (error) {
+    console.log("Email not sent", error);
+  }
 };
